requirements.txt
Package                      Version
---------------------------- -----------
altair                       5.4.1
annotated-types              0.7.0
anyio                        4.6.2.post1
attrs                        24.2.0
beautifulsoup4               4.12.3
blinker                      1.9.0
cachetools                   5.5.0
certifi                      2024.8.30
cffi                         1.17.1
charset-normalizer           3.4.0
click                        8.1.7
clickhouse-connect           0.8.6
colorama                     0.4.6
distro                       1.9.0
exceptiongroup               1.2.2
gemini-api                   0.1.6
gitdb                        4.0.11
GitPython                    3.1.43
google-ai-generativelanguage 0.6.10
google-api-core              2.23.0
google-api-python-client     2.152.0
google-auth                  2.36.0
google-auth-httplib2         0.2.0
google-generativeai          0.8.3
googleapis-common-protos     1.66.0
greenlet                     3.1.1
grpcio                       1.67.1
h11                          0.14.0
httpcore                     1.0.6
httplib2                     0.22.0
httpx                        0.27.2
idna                         3.10
Jinja2                       3.1.4
jiter                        0.7.0
jsonpatch                    1.33
jsonpointer                  3.0.0
jsonschema                   4.23.0
jsonschema-specifications    2024.10.1
langchain-core               0.3.17
langchain-text-splitters     0.3.2
langsmith                    0.1.142
lz4                          4.3.3
markdown-it-py               3.0.0
MarkupSafe                   3.0.2
mdurl                        0.1.2
narwhals                     1.13.4
numpy                        1.26.4
openai                       1.54.3
orjson                       3.10.11
outcome                      1.3.0.post0
packaging                    24.2
pandas                       2.2.3
pillow                       11.0.0
pinecone-client              5.0.1
pinecone-plugin-inference    1.1.0
pinecone-plugin-interface    0.0.7
pip                          24.3.1
proto-plus                   1.25.0
protobuf                     5.28.3
pyarrow                      18.0.0
pyasn1                       0.6.1
pyasn1_modules               0.4.1
pycparser                    2.22
pydantic                     2.9.2
pydantic_core                2.23.4
pydeck                       0.9.1
Pygments                     2.18.0
pyparsing                    3.2.0
PySocks                      1.7.1
python-dateutil              2.9.0.post0
python-dotenv                1.0.1
pytz                         2024.2
PyYAML                       6.0.2
referencing                  0.35.1
requests                     2.32.3
requests-toolbelt            1.0.0
rich                         13.9.4
rpds-py                      0.21.0
rsa                          4.9
scipy                        1.13.1
selenium                     4.26.1
setuptools                   49.2.1
six                          1.16.0
smmap                        5.0.1
sniffio                      1.3.1
sortedcontainers             2.4.0
soupsieve                    2.6
SQLAlchemy                   2.0.36
streamlit                    1.40.1
tenacity                     9.0.0
toml                         0.10.2
tornado                      6.4.1
tqdm                         4.67.0
trio                         0.27.0
trio-websocket               0.11.1
typing_extensions            4.12.2
tzdata                       2024.2
uritemplate                  4.1.1
urllib3                      2.2.3
watchdog                     6.0.0
websocket-client             1.8.0
wsproto                      1.2.0
zstandard                    0.23.0
